Fix picking of spawn eggs not working on fabric
The addModdedEggs hack usually only needed for forge now also works for fabric and enables me to use the single ModSpawnEggItem class in common.

Drawback of that is, that the accessor seems to break hotcode reloading, so gotta comment that for intenstive testing or investigate how to only recompile, transform and hotreload a  single class with architectury.